@@ -1,10 +1,11 @@
 package no.fint.sse;
 
-import javax.annotation.Priority;
-import javax.validation.constraints.NotNull;
-import javax.ws.rs.Priorities;
-import javax.ws.rs.client.ClientRequestContext;
-import javax.ws.rs.client.ClientRequestFilter;
+import jakarta.annotation.Priority;
+import jakarta.validation.constraints.NotNull;
+import jakarta.ws.rs.Priorities;
+import jakarta.ws.rs.client.ClientRequestContext;
+import jakarta.ws.rs.client.ClientRequestFilter;
+
 import java.io.IOException;
 
 @Priority(Priorities.HEADER_DECORATOR)
